% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.IAMCTimeseries.R
\name{as.IAMCTimeseries}
\alias{as.IAMCTimeseries}
\title{Converts an IAMC long format into a wide format}
\usage{
as.IAMCTimeseries(df, fill = NA)
}
\arguments{
\item{df}{A fill value. Defaults to \code{NA}}
}
\description{
Pivots periods to columns so that values are given as a time series table rather than a long list.
}
\examples{
write.IAMCxlsx(quitte_example_data, tempfile())

}
\author{
Michaja Pehl, Tonn Rüter
}
