% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_addScenChange.R
\name{calc_addScenIncrease}
\alias{calc_addScenIncrease}
\alias{calc_addScenRatio}
\alias{calc_addScenPercentChange}
\title{Add Scenario Change}
\usage{
calc_addScenIncrease(
  d,
  variable,
  reference_scenario,
  reference_short_name = reference_scenario
)

calc_addScenRatio(
  d,
  variable,
  reference_scenario,
  reference_short_name = reference_scenario
)

calc_addScenPercentChange(
  d,
  variable,
  reference_scenario,
  reference_short_name = reference_scenario
)
}
\arguments{
\item{d}{A quitte-like data frame.}

\item{variable}{A vector of variable names.}

\item{reference_scenario}{The scenario changes are calculated to.}

\item{reference_short_name}{Short name of reference scenario (appended to the
variable name).}
}
\value{
A quitte-like data frame with the calculated changes.
}
\description{
These functions add the absolute increase over, the ratio relative to, or the
percent change relative to a reference scenario.  Variable names have
\verb{|Int over <reference_short_name>}, \verb{|Ratio to <reference_short_name>}, or
\verb{|\% Change to <reference_short_name>} appended, units are either unmodified,
or set to \code{ratio} or \verb{\%}, respectively.
}
\examples{
quitte_example_data |>
    calc_addScenIncrease(c('Consumption', 'GDP|PPP'),
                         'r7552c_REF_Def05-rem-5', 'REF_Def05')

}
