% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_growthrate.R
\name{calc_growthrate}
\alias{calc_growthrate}
\title{Calculates the growth rate in '\%/yr' for variables}
\usage{
calc_growthrate(x, only.new = FALSE, filter.function = identity)
}
\arguments{
\item{x}{anything with an as.quitte method (data.frame, quitte or magclass object, mif file, ...)}

\item{only.new}{If \code{FALSE} (the default), add new variables to existing
ones.  If \code{TRUE}, return only new variables.}

\item{filter.function}{A function used to filter data before calculating growth rates.
If instead a character vector is passed, only variables matching this vector are used.}
}
\value{
data as a quitte object
}
\description{
Calculates the growth rate in '\%/yr' for variables
}
\details{
If, for example, your data contains the data in 2070 and 2060, the growth rate returned
for 2070 is calculated as 100 * ((d2070/d2060)^(1/10) - 1).
No growth rate can be calculated for the first year of the data.
Infinite or undefined values (for example if d2060 = 0) are dropped.
}
\examples{
\dontrun{
GDPgrowth <- calc_growthrate(quitte_example_data, only.new = TRUE, filter.function = "GDP|PPP")
alldata <- calc_growthrate(quitte_example_data, only.new = FALSE,
                           filter.function = function(x) filter(x, grepl("GDP", .data$variable)))
}

}
\author{
Oliver Richters
}
