% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_mode.R
\name{calc_mode}
\alias{calc_mode}
\title{Calculate the mode of a sample}
\usage{
calc_mode(v)
}
\arguments{
\item{v}{A vector.}
}
\value{
The mode, or a vector of modes if the sample is multi-modal.
}
\description{
Calculate the mode of a sample
}
\examples{
calc_mode(c(1, 1, 100))
calc_mode(c('a', 'a', 'b', 'c', 'c'))
}
\author{
Michaja Pehl
}
