% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMitigationCost.R
\name{calcMitigationCost}
\alias{calcMitigationCost}
\title{Calculate mitigation costs}
\usage{
calcMitigationCost(data, scenBau, scenPol, yearFrom = 2010,
  yearTo = 2100, nameVar = "Consumption",
  nameDisrate = "Interest Rate t/(t-1)|Real", discount = 0.05)
}
\arguments{
\item{data}{quitte object}

\item{scenBau}{baseline scenario name}

\item{scenPol}{policy scenario name}

\item{yearFrom}{the startyear}

\item{yearTo}{the endyear}

\item{nameVar}{Name of the variable containing consumption. Defaults to
"Consumption"}

\item{nameDisrate}{Name of the variable for the discount rate, only needed
if discount=endo. Defaults to "Interest Rate t/(t-1)|Real"}

\item{discount}{discount rate - exogenous only for now}
}
\value{
regional mitigation costs (quitte object)
}
\description{
Calculate mitigation costs
}
\examples{

  \dontrun{
    calcMitigationCost(qd,"BAU","POL")
  }

}
\author{
Anselm Schultes
}
