% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartesian.R
\name{cartesian}
\alias{cartesian}
\title{Generate cartesian product from to character vectors}
\usage{
cartesian(x, y, sep = ".")
}
\arguments{
\item{x, y}{objects that can be coerced to character vectors}

\item{sep}{a character string that will seperate the elements of \code{x}
and \code{y} in the output. Defaults to \code{'.'}.}
}
\value{
A character vector of the concatenated elements of \code{x} and
\code{y}.
}
\description{
Generate cartesian product from to character vectors
}
\examples{
cartesian(c('a', 'b'), 1:3)
# [1] "a.1" "a.2" "a.3" "b.1" "b.2" "b.3"

}
