% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df.2.named.vector.R
\name{df.2.named.vector}
\alias{df.2.named.vector}
\title{Data frame to named vector}
\usage{
df.2.named.vector(.data)
}
\arguments{
\item{.data}{A data frame with at least two columns.}
}
\value{
A named vector.
}
\description{
Turns the two first columns of a data frame into a named vector, taking the
values from the second and the names from the first column.
}
\examples{
data <- data.frame(names = c("one", "two", "three"), values = 1:3)
data
df.2.named.vector(data)
}
\author{
Michaja Pehl
}
