% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_variation.R
\name{df_variation}
\alias{df_variation}
\title{Data Frame Variation}
\usage{
df_variation(x)
}
\arguments{
\item{x}{A data frame.}
}
\value{
A data frame.
}
\description{
Removes all columns from a data frame which have only identical data, to
facilitate a quick overview.
}
\examples{
(quitte_example_data['Consumption' == quitte_example_data$variable,] -> x)

df_variation(x)
}
