% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/escape_cntrl.R
\name{escape_cntrl}
\alias{escape_cntrl}
\title{Escape Control Characters}
\usage{
escape_cntrl(x, escape_quotes = c("none", "single", "double", "both"))
}
\arguments{
\item{x}{A character vector.}

\item{escape_quotes}{Should quotes be escaped?  One of \code{none}, \code{single},
\code{double}, or \code{both}.}
}
\value{
A character vector with escaped control characters.
}
\description{
Escapes control characters with back slashes, such that they are safe to
print.  Escaped characters are:\tabular{rcl}{
   code \tab symbol \tab description \cr
   \verb{\\0x07} \tab \verb{\\a} \tab bell \cr
   \verb{\\0x08} \tab \verb{\\b} \tab backspace \cr
   \verb{\\0x09} \tab \verb{\\t} \tab horizontal tab \cr
   \verb{\\0x0a} \tab \verb{\\n} \tab line feed \cr
   \verb{\\0x0b} \tab \verb{\\v} \tab vertical tab \cr
   \verb{\\0x0c} \tab \verb{\\f} \tab form feed \cr
   \verb{\\0x0d} \tab \verb{\\r} \tab carriage return \cr
   \verb{\\0x22} \tab \verb{\\"} \tab double quote \cr
   \verb{\\0x27} \tab \verb{\\'} \tab single quote \cr
   \verb{\\0x5c} \tab \verb{\\\\} \tab back slash \cr
}
}
\examples{
escape_cntrl(c('foo\nbar\t\tbazz', 'foo\abar'))
}
