% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor.data.frame.R
\name{factor.data.frame}
\alias{factor.data.frame}
\title{(Re-) Factorise factor and character columns in data frame}
\usage{
factor.data.frame(df)
}
\arguments{
\item{df}{a data frame}
}
\value{
a data frame
}
\description{
\code{factor.data.frame()} turns character columns in a data frame into
factor columns and refactorises factor columns, silently dropping unused
levels.
}
\examples{
df <- data.frame(
    character = letters[1:5],
    factor = as.factor(LETTERS[1:5]),
    value = 1:5,
    stringsAsFactors = FALSE)
df <- df[1:3,]
str(df)
str(factor.data.frame(df))
}
\author{
Michaja Pehl
}

