% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorise.R
\name{factorise}
\alias{factorise}
\title{Factorise}
\usage{
factorise(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
A factor from \code{x}, with levels in the same order as they appear
    in within \code{x}.
}
\description{
Generate a factor with levels in prescribed order.
}
\examples{
factor(c('a', 'c', 'b'))
factorise(c('a', 'c', 'b'))
}
\author{
Michaja Pehl
}
