% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatherIAMCTimeseries.R
\name{gatherIAMCTimeseries}
\alias{gatherIAMCTimeseries}
\title{gatherIAMCTimeseries}
\usage{
gatherIAMCTimeseries(df, quitte_case = FALSE)
}
\arguments{
\item{df}{\code{quitte} data frame in IAMC wide format (i.e. years/periods are columns)}

\item{quitte_case}{In IAMC case columns first letters are capitalized, quitte case is lower case. Default is FALSE}
}
\value{
\code{quitte} data frame in long IAMC (= \code{quitte}) format
}
\description{
Pivots periods columns (as they occur in IAMC wide format/timeseries) to the typical \code{quitte} long list format
}
\examples{
\dontrun{
df <- quitte_example_data \%>\% as.IAMCTimeseries()
df \%>\% gatherIAMCTimeseries()
}
}
\author{
Tonn Rüter
}
