% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getColValues.R
\name{getColValues}
\alias{getColValues}
\title{Retrieves values from one column of a quitte object}
\usage{
getColValues(df, colVar)
}
\arguments{
\item{df}{quitte object}

\item{colVar}{name of the column of interest}
}
\value{
a  vector containing the values of the column without duplicates. The
class of the returned vector is either numeric or character
}
\description{
Retrieves values from one column of a quitte object
}
\examples{
data(mtcars)
getColValues(mtcars,"mpg")
}
\author{
Antoine Levesque
}
