% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getModels.R
\name{getModels}
\alias{getModels}
\title{Retrieves models from a quitte object}
\usage{
getModels(df)
}
\arguments{
\item{df}{quitte object}
}
\value{
a character vector containing the values of the column without
duplicates
}
\description{
Retrieves models from a quitte object
}
\examples{

  \dontrun{
    getPeriods(dataframe)
  }
}
\author{
Antoine Levesque
}
