% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPeriods.R
\name{getPeriods}
\alias{getPeriods}
\title{Retrieves periods from a quitte object}
\usage{
getPeriods(df)
}
\arguments{
\item{df}{quitte object}
}
\value{
a numeric vector containing the values of the column without
duplicates
}
\description{
Retrieves periods from a quitte object
}
\examples{

  \dontrun{
    getPeriods(dataframe)
  }
}
\author{
Antoine Levesque
}
