% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRegs.R
\name{getRegs}
\alias{getRegs}
\title{Retrieves regions from a quitte object}
\usage{
getRegs(df)
}
\arguments{
\item{df}{quitte object}
}
\value{
a character vector containing the values of the column without
duplicates
}
\description{
Retrieves regions from a quitte object
}
\examples{

  \dontrun{
    getRegs(dataframe)
  }
}
\author{
Antoine Levesque
}
