% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getScenarios.R
\name{getScenarios}
\alias{getScenarios}
\title{Retrieves scenarios from a quitte object}
\usage{
getScenarios(df)
}
\arguments{
\item{df}{quitte object}
}
\value{
a character vector containing the values of the column without duplicates
}
\description{
Retrieves scenarios from a quitte object
}
\examples{

  \dontrun{
    getPeriods(dataframe)
  }
}
\author{
Antoine Levesque
}
