% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVars.R
\name{getVars}
\alias{getVars}
\title{Retrieves variable names from a quitte object}
\usage{
getVars(df)
}
\arguments{
\item{df}{quitte object}
}
\value{
a character vector containing the values of the column without
duplicates
}
\description{
Retrieves variable names from a quitte object
}
\examples{

  \dontrun{
    getVars(dataframe)
  }
}
\author{
Antoine Levesque
}
