% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_colour_hue.R
\name{gg_colour_hue}
\alias{gg_colour_hue}
\title{Get n Colours}
\usage{
gg_colour_hue(n)
}
\arguments{
\item{n}{Either the number of colours to generate, or a character vector
which will be used as names for the returned colours.}
}
\value{
A vector of character strings which can be used as color
specifications by R graphics functions.  The vector is named if n is a
character vector.
}
\description{
Get n colours, evenly spaced along the colour wheel.  Just like the ones
\code{\link[ggplot2:scale_colour_hue]{ggplot2::scale_colour_hue()}} is using.
}
\examples{
gg_colour_hue(5)
gg_colour_hue(letters[1:3])

}
