% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthrate.R
\name{growthrate}
\alias{growthrate}
\title{Calculates the growth rate of all variables in '\%/yr' and returns a sorted data.frame
where each variable name gets '|Growth rate' appended.
If, for example, your data contains the data in 2070 and 2060, the growth rate returned
for 2070 is calculated as 100 * ((d2070/d2060)^(1/10) - 1).
No growth rate can be calculated for the first year of the data.
Infinite or undefined values (for example if d2060 = 0) are dropped.}
\usage{
growthrate(x)
}
\arguments{
\item{x}{A quitte object}
}
\value{
the sorted quitte object
}
\description{
Calculates the growth rate of all variables in '\%/yr' and returns a sorted data.frame
where each variable name gets '|Growth rate' appended.
If, for example, your data contains the data in 2070 and 2060, the growth rate returned
for 2070 is calculated as 100 * ((d2070/d2060)^(1/10) - 1).
No growth rate can be calculated for the first year of the data.
Infinite or undefined values (for example if d2060 = 0) are dropped.
}
\author{
Oliver Richters
}
