% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline.data.frame.R
\name{inline.data.frame}
\alias{inline.data.frame}
\title{Convert a vector of stings to a data frame}
\usage{
inline.data.frame(table, sep = ";", quote = "")
}
\arguments{
\item{table}{vector of strings}

\item{sep}{item separator within strings}

\item{quote}{quote character for masking separators}
}
\value{
a data frame
}
\description{
\code{inline.data.frame()} converts a vector of strings that contain
separated items into a data frame.
}
\examples{
inline.data.frame(c("letters; numbers", "A; 1", "B; 2"))

}
\author{
Michaja Pehl
}

