% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_to_data_frame.R
\name{list_to_data_frame}
\alias{list_to_data_frame}
\alias{list_to_data_frame_}
\title{Convert mapping list to data frame}
\usage{
list_to_data_frame(l, ...)

list_to_data_frame_(l, category = "category", item = "item")
}
\arguments{
\item{l}{A named list of character vectors.}

\item{...}{Unquoted names of category and item columns. Defaults to
'category' and 'item'.}

\item{category}{Name of category column. Defaults to 'category'.}

\item{item}{Name of item column. Defaults to 'item'.}
}
\value{
A data frame.
}
\description{
Convert mapping list to data frame
}
\examples{
l <- list(Africa = c('Egypt', 'Tanzania'),
          Europe = c('Portugal', 'Ukraine', 'Denmark'))
list_to_data_frame(l, region, country)

}
\author{
Michaja Pehl
}
