% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/looks_like_quitte.R
\name{looks_like_quitte}
\alias{looks_like_quitte}
\title{Does a data frame match the quitte definition?}
\usage{
looks_like_quitte(df, verbose = NULL)
}
\arguments{
\item{df}{An object to test for \code{quitte} structure.}

\item{verbose}{Either \code{NULL}, or one of \code{'abort'}, \code{'warn'}, or \verb{'inform}',
to report if any of the criteria of the quitte definition are not
fulfilled using \verb{[cli::cli_abort]}, \verb{[cli::cli_warn]}, or
\verb{[cli::cli_inform]}.}
}
\value{
\code{TRUE} if all criteria are met, \code{FALSE} otherwise.
}
\description{
Checks if \code{df} fulfills all criteria of a "quitte data frame", namely:
\itemize{
\item Must be of class "data.frame".
\item Must have the mandatory columns "model", "scenario", "region", "variable",
"unit", "period", and "value".
\item The "period" column must be of class "integer" or "POSIXct".
\item The "value" column must be of type "numeric".
}
}
\examples{
library(dplyr)

quitte_example_data \%>\%
    looks_like_quitte()

quitte_example_data \%>\%
    select(-'model') \%>\%
    mutate(period = as.character(period)) \%>\%
    looks_like_quitte(verbose = 'inform')

}
