% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.ISOyear.R
\name{make.ISOyear}
\alias{make.ISOyear}
\title{speedily converting years to POSIXct values}
\usage{
make.ISOyear(years)
}
\arguments{
\item{years}{ignored.}
}
\value{
The \code{\link[=ISOyear]{ISOyear()}} function.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\details{
This function was deprecated because the \code{\link[=ISOyear]{ISOyear()}} function can be used
directly.
}
\examples{
ISOyear <- make.ISOyear()
ISOyear(c(2005, 2010, 2100, 1900))
# ->
ISOyear(c(2005, 2010, 2100, 1900))

}
\keyword{internal}
