% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mif2rds.R
\name{mif2rds}
\alias{mif2rds}
\title{MIF2RDS converter}
\usage{
mif2rds(input, output = "output.rds")
}
\arguments{
\item{input}{Path to the MIF file to be converted}

\item{output}{File name of the output file}
}
\description{
Converts mif file into more memory efficient rds file. Additionally,
removes NA and duplicate entries.
}
\author{
Jan Philipp Dietrich
}
