% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_text.R
\name{mutate_text}
\alias{mutate_text}
\title{Uses mutate based on a character vector#'}
\usage{
mutate_text(df, s)
}
\arguments{
\item{df}{a data frame}

\item{s}{a character string containing the formula to be applied in mutate}
}
\value{
A data frame transformed with the mutate function
}
\description{
Uses mutate based on a character vector#'
}
\examples{
df = data.frame(x = c(1,2),
                y = c(3,4))
form = "z = x + y"
mutate_text(df,form)                 

}
\author{
Antoine Levesque
}
