% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order.levels.R
\name{order.levels}
\alias{order.levels}
\title{Order data frame factor levels}
\usage{
order.levels(df, ..., drop.extra.levels = TRUE)
}
\arguments{
\item{df}{a data frame}

\item{...}{name-value pairs assigning level order to factors}

\item{drop.extra.levels}{if \code{TRUE} (default) levels not present in the
factor are silently dropped}
}
\value{
a data frame
}
\description{
Order data frame factor levels
}
\examples{
require(dplyr)
df <- data.frame(name = letters[1:3], value = 1:3)
df \%>\% getElement("name")
df \%>\% order.levels(name = c("c", "a", "b", "x")) \%>\% getElement("name")
}
\author{
Michaja Pehl
}

