% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_periods.R
\name{pivot_periods}
\alias{pivot_periods}
\alias{pivot_periods_wider}
\alias{pivot_periods_longer}
\title{Pivot Data Frame on Periods}
\usage{
pivot_periods(df, direction = NULL, fill = NA, drop_na = TRUE)

pivot_periods_wider(df, fill = NA)

pivot_periods_longer(df, drop_na = TRUE)
}
\arguments{
\item{df}{A data frame.}

\item{direction}{Direction in which to pivot.  Either \code{'wider'} or
\code{'longer'}.  Determined automatically for the default \code{NULL}.}

\item{fill}{A scalar value that fills in missing values when pivoting wider.
Defaults to \code{NA}.}

\item{drop_na}{If \code{TRUE} (the default), will drop rows that contain only
\code{NA}s in the \code{value} column when pivoting longer.}
}
\value{
The pivoted data frame.
}
\description{
Pivot data frame on periods either to wide or long format.
}
\examples{
library(dplyr)

(x <- quitte_example_data \%>\%
        filter(first(scenario) == scenario,
               first(region) == region,
               variable \%in\% head(unique(variable), n = 7),
               period <= 2030) \%>\%
        slice_sample(prop = 0.9) \%>\%
        pivot_periods_wider())

x \%>\%
    pivot_periods_longer()

}
\seealso{
\code{\link[tidyr:pivot_longer]{pivot_longer()}},
\code{\link[tidyr:pivot_wider]{pivot_wider()}}.
}
