% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepQuitteForScatter.R
\name{prepQuitteForScatter}
\alias{prepQuitteForScatter}
\title{selects two variables from a long format Quitte and puts them into a wide
format Quitte}
\usage{
prepQuitteForScatter(quitteIn, varNames)
}
\arguments{
\item{quitteIn}{Quitte with original data}

\item{varNames}{Vector with two variable names that must be contained in
QuitteIn$variable}
}
\value{
quitte object
}
\description{
QuitteIn contains two variables containedin varNames that should be plotted
in a scatter plot. The functions forms the new QuitteOut with the variables
x and y. QuitteOut can be used in ggplot with plotting x and y. The unit
needs to be replaced by None.
}
\examples{

  \dontrun{
    quitteOut <- prepQuitteForScatter(quitteIn, c('Emissions|CO2', 'Price|Carbon'))
  }
}
\author{
Nico Bauer, Anselm Schultes, Jerome Hilaire
}
