% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.quitte.R, R/is.quitte.R
\docType{class}
\name{quitte-class}
\alias{quitte-class}
\alias{as.quitte}
\alias{as.quitte-methods}
\alias{as.quitte,character-method}
\alias{is.quitte}
\title{Class "quitte" ~~~}
\usage{
as.quitte(x, periodClass = "integer", addNA = FALSE, na.rm = FALSE)

is.quitte(x, warn = TRUE)
}
\arguments{
\item{x}{An object that should be either tested or transformed as/to an
quitte-object.}

\item{periodClass}{integer or POSIXct}

\item{addNA}{modifies a factor by turning NA into an extra level (so that NA values are counted in tables, for instance).}

\item{na.rm}{if set to TRUE entries with value NA will be removed}

\item{warn}{display warnings or not}
}
\description{
The quitte class is a more standardized data.frame format. \code{is.quitte}
tests if \code{x} is an quitte-object, \code{as.quitte} transforms \code{x}
to an quitte-object (if possible).
}
\section{Functions}{
\itemize{
\item \code{is.quitte}: quitte check
}}

\author{
Jan Philipp Dietrich
}
\keyword{classes}
