% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quitte-package.R
\docType{package}
\name{quitte-package}
\alias{quitte-package}
\alias{_PACKAGE}
\alias{quitte}
\title{The quitte R package}
\description{
Bits and pieces for easier handling of quitte-style data frames.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pik-piam.github.io/quitte/}
}

}
\author{
Michaja Pehl
}
