% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quitte2quantiles.R
\name{quitte2quantiles}
\alias{quitte2quantiles}
\title{calculate quantiles}
\usage{
quitte2quantiles(x, probs = c(q0 = 0, q25 = 0.25, q50 = 0.5, q75 = 0.75,
  q100 = 1), grouping = c("region", "variable", "period", "scenario"))
}
\arguments{
\item{x}{dataframe to add quantiles}

\item{probs}{default=c(q0=0,q25=0.25,q50=0.5,q75=0.75,q100=1)}

\item{grouping}{default=c("region", "variable", "period", "scenario")}
}
\description{
calculate quantiles
}
\examples{

  \dontrun{
    p <- x.minmax = quitte2quantiles(x,probs=c("min"=0,"max"=1))
  }

}
\author{
Gunnar Luderer, Lavinia Baumstark
}
