% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quitte_example_data.R
\docType{data}
\name{quitte_example_data}
\alias{quitte_example_data}
\title{quitte example data}
\usage{
quitte_example_data
}
\description{
A quitte object with some example data from the REMIND model.
}
\author{
Michaja Pehl
}
