% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quitte_example_dataAR6.R
\docType{data}
\name{quitte_example_dataAR6}
\alias{quitte_example_dataAR6}
\title{quitte example data with three models (REMIND, GCAM, MESSAGEix) and two
scenarios (Current Policies, Delayed transition)}
\usage{
quitte_example_dataAR6
}
\description{
A quitte object with some example data from NGFS Phase 3.
}
\author{
Oliver Richters
}
