% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quitteSort.R
\name{quitteSort}
\alias{quitteSort}
\title{Sorts a quitte object in a standardized way:
model -> scenario -> variable -> unit -> region -> period}
\usage{
quitteSort(x)
}
\arguments{
\item{x}{anything with an as.quitte method (data.frame, quitte or magclass object, mif file, ...)}
}
\value{
the sorted quitte object
}
\description{
Sorts a quitte object in a standardized way:
model -> scenario -> variable -> unit -> region -> period
}
\author{
Oliver Richters
}
