% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.filter.snapshot.R
\name{read.filter.snapshot}
\alias{read.filter.snapshot}
\title{Reads IAMC-style .csv files obtained as a IIASA snapshot into a quitte data frame,
filtering the data. This function is helpful if the csv file is large and R runs out
of memory loading it completely.}
\usage{
read.filter.snapshot(file, keep = list())
}
\arguments{
\item{file}{Path of single IAMC-style .csv/.mif file}

\item{keep}{list with quitte columns as names and data points that should be kept.}
}
\value{
A quitte data frame.
}
\description{
Reads IAMC-style .csv files obtained as a IIASA snapshot into a quitte data frame,
filtering the data. This function is helpful if the csv file is large and R runs out
of memory loading it completely.
}
\examples{
\dontrun{
read.filter.snapshot("snapshot.csv", list(scenario = c("CurPol", "NDC"), region = "World"))
}

}
\author{
Oliver Richters
}
