% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.quitte.R
\name{read.quitte}
\alias{read.quitte}
\title{read IAMC-style .csv files into quitte data frame}
\usage{
read.quitte(file, sep = ";", quote = "",
            na.strings = c("UNDF", "NA", "N/A", "n_a"),
            convert.periods = FALSE)
}
\arguments{
\item{file}{IAMC-style .csv file}

\item{sep}{column separator; defaults to ";"}

\item{quote}{quote characters; empty by default}

\item{na.strings}{entries to interpret as NA; defaults to
c("UNDF", "NA", "N/A", "n_a")}

\item{convert.periods}{if TRUE, periods are converted to POSIXct, defaults
to FALSE}
}
\value{
a quitte data frame
}
\description{
read IAMC-style .csv files into quitte data frame
}
\examples{
\dontrun{
mif <- read.quitte("some/data/file.mif")
csv <- read.quitte("some/data/file.csv", sep = ",", quote = '"')
}

}
\author{
Michaja Pehl
}

