% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mif_header.R
\name{read_mif_header}
\alias{read_mif_header}
\title{Read .mif Header}
\usage{
read_mif_header(file, sep = ";", comment = "#")
}
\arguments{
\item{file}{A path to a \code{.mif} file.}

\item{sep}{Column separator, defaults to ";".}

\item{comment}{A character which at line start signifies the optional comment
header with metadata at the head of \code{file}, defaults to "#".}
}
\value{
A \code{list} with elements \code{header}, \code{comment_header}, and
\code{useless.last.column}.
}
\description{
Read .mif Header
}
\author{
Michaja Pehl
}
