% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remind_timesteps-data.R
\docType{data}
\name{remind_timesteps}
\alias{remind_timesteps}
\title{REMIND time steps}
\usage{
remind_timesteps
}
\description{
A data frame containing the weights with which years contribute
to specific periods.
}
\examples{
require(tidyverse)

remind_timesteps \%>\%
    filter(period \%in\% c(2055, 2060, 2070)) \%>\%
    spread(period, weight, fill = 0) \%>\%
    print(n = Inf)

}
\author{
Michaja Pehl
}
