% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportDuplicates.R
\name{reportDuplicates}
\alias{reportDuplicates}
\title{Finds duplicates in a quitte object and warns about them.
It removes any |+| notation and also warns if units or value differ}
\usage{
reportDuplicates(mifdata)
}
\arguments{
\item{mifdata}{object that can be converted with as.quitte into quitte data.frame}
}
\description{
Finds duplicates in a quitte object and warns about them.
It removes any |+| notation and also warns if units or value differ
}
