% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setXor.R
\name{setXor}
\alias{setXor}
\title{Set Xor}
\usage{
setXor(x, y)
}
\arguments{
\item{x, y}{Objects to perform set function on.}
}
\description{
Performs \eqn{(x \cup y) \setminus (x \cap y)} on
parameters, returning all elements that are in either x or y, but not both.
}
\examples{
x <- c('a', 'b', 'c')
y <- c('b', 'c', 'd')
setXor(x, y)

}
\author{
Michaja Pehl
}
