% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signif_frac.R
\name{signif_frac}
\alias{signif_frac}
\title{Round fractions only}
\usage{
signif_frac(x, digits = 3)
}
\arguments{
\item{x}{A numeric vector.}

\item{digits}{Number of significant digits, defaults to 3.}
}
\value{
A numeric vector.
}
\description{
Round numbers to significant digits, preserving integer parts.
}
\examples{
set.seed(0)
x <- runif(8, -1, 1) * 10 ^ seq.int(5, -2)

data.frame(number           = x,
           formatted        = sprintf('\%g', x),
           `w/ signif`      = sprintf('\%g', signif(x, 3)),
           `w/ signif_frac` = sprintf('\%g', signif_frac(x)),
           check.names = FALSE)

}
