% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signrange.R
\name{signrange}
\alias{signrange}
\title{Sign Range}
\usage{
signrange(x, na.rm = TRUE)
}
\arguments{
\item{x}{A numerical vector.}

\item{na.rm}{Should \code{NA}s be ignored?}
}
\value{
A character string of signs found in \code{x}.
}
\description{
Returns the range of signs in a numerical vector as a character string.
}
\examples{
signrange(-1)
signrange(0)
signrange(1)
signrange(c(-1, 0))
signrange(c(0, 1))
signrange(c(-1, 1))
}
\author{
Michaja Pehl
}
