% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_total.R
\name{sum_total}
\alias{sum_total}
\title{Sum over one dimension of a data frame}
\usage{
sum_total(data, group, value = NA, name = "Total")
}
\arguments{
\item{data}{a data frame}

\item{group}{column for which the sum is to be calculated}

\item{value}{column of the numbers to be summed}

\item{name}{entry in column \code{group} for the sum; defaults to
\code{"Total"}}
}
\value{
a data frame
}
\description{
\code{sum_total()} is a short-hand function to calculate and insert the sum
of a category in a data frame.
}
\examples{
require(dplyr)
data <- expand.grid(UPPER = LETTERS[1:2], lower = letters[24:26],
                    numbers = 1:2) \%>\%
    mutate(value = 1:n()) \%>\%
    group_by(UPPER)
data
data \%>\%
    sum_total(lower)

}
\author{
Michaja Pehl
}

