% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolExtractSortScaleQuitte.R
\name{toolExtractSortScaleQuitte}
\alias{toolExtractSortScaleQuitte}
\title{prepare data for plots}
\usage{
toolExtractSortScaleQuitte(x, scen, vars, var.scaling = 1,
  regi = c("World"), prd = getPeriods(x))
}
\arguments{
\item{x}{dataframe to prepare}

\item{scen}{scenario to select}

\item{vars}{variables to select}

\item{var.scaling}{scaling of the variables, default=1}

\item{regi}{region to select, default="World"}

\item{prd}{period to select, default=getPeriods(x)}
}
\description{
prepare data for plots
}
\examples{

  \dontrun{
    p <- toolExtractSortScaleQuitte(x,scen=c("BAU"),
                                   vars=c("Emi|CO2","FE|Industry"),
                                   regi=c("EUR","LAM"),prd=c(2005,2030,2050))
  }



}
\author{
Gunnar Luderer, Lavinia Baumstark
}
