% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique_or_levels.R
\name{unique_or_levels}
\alias{unique_or_levels}
\title{Unique or levels}
\usage{
unique_or_levels(x)
}
\arguments{
\item{x}{A character vector or a factor.}
}
\value{
A character vector with the unique elements of \code{x} if it is a
character vector, or the levels of \code{x} if it is a factor.
}
\description{
Abstract the differences between character vectors and factors.
}
\author{
Michaja Pehl
}
