% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unquitte.R
\name{unquitte}
\alias{unquitte}
\title{Remove quitte class atribute}
\usage{
unquitte(x)
}
\arguments{
\item{x}{Any object.}
}
\value{
\code{x}, with the \code{quitte} class attribute removed (if any).
}
\description{
Remove quitte class atribute
}
\examples{
quitte_example_data |> class()
quitte_example_data |> unquitte() |> class()
mtcars |> class()
mtcars |> unquitte() |> class()
}
