% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useWorld.R
\name{useWorld}
\alias{useWorld}
\title{useWorld}
\usage{
useWorld(qu)
}
\arguments{
\item{qu}{A quitte object with a global region.}
}
\value{
The quitte object with the global region named "World".
}
\description{
Changes the identifier for the global region to "World".
}
\author{
Patrick Rein
}
