% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_growthrate.R
\name{calc_growthrate}
\alias{calc_growthrate}
\title{Calculates the growth rate of all variables in '\%/yr' and returns a quitte data.frame
where each variable name gets \verb{ [Growth Rate]} appended.}
\usage{
calc_growthrate(x)
}
\arguments{
\item{x}{anything with an as.quitte method (data.frame, quitte or magclass object, mif file, ...)}
}
\value{
the sorted quitte object
}
\description{
Calculates the growth rate of all variables in '\%/yr' and returns a quitte data.frame
where each variable name gets \verb{ [Growth Rate]} appended.
}
\details{
If, for example, your data contains the data in 2070 and 2060, the growth rate returned
for 2070 is calculated as 100 * ((d2070/d2060)^(1/10) - 1).
No growth rate can be calculated for the first year of the data.
Infinite or undefined values (for example if d2060 = 0) are dropped.
}
\author{
Oliver Richters
}
