% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_quantiles.R
\name{calc_quantiles}
\alias{calc_quantiles}
\alias{calc_quantiles_}
\title{Sample Quantiles}
\usage{
calc_quantiles(
  .data,
  value = NA,
  probs = c(q0 = 0, q25 = 0.25, q50 = 0.5, q75 = 0.75, q100 = 1),
  na.rm = TRUE,
  type = 7
)

calc_quantiles_(
  .data,
  value = "value",
  probs = c(q0 = 0, q25 = 0.25, q50 = 0.5, q75 = 0.75, q100 = 1),
  na.rm = TRUE,
  type = 7
)
}
\arguments{
\item{.data}{a data frame, possibly grouped}

\item{value}{column name for which sample quantiles should be calculated}

\item{probs}{named numeric vector of probabilities with values in
\eqn{[0, 1]}.}

\item{na.rm}{logical; if true, any \code{\link{NA}} and \code{\link{NaN}}s
are removed from x before the quantiles are computed.}

\item{type}{an integer between 1 and 9 select one of the nine quantile.
algorithms detailed in \code{\link[stats]{quantile}} to be used.}
}
\value{
a data frame.
}
\description{
This is a wrapper function for \code{\link[stats]{quantile}} or easy use with
data frames.
}
\examples{
require(dplyr)
require(tidyr)
data <- tibble(group = rep(c("A", "B"), 10),
                   value = 1:20) \%>\%
                   arrange(group) \%>\%
                   group_by(group)

data \%>\%
 calc_quantiles() \%>\%
 spread(quantile, value)

}
\author{
Michaja Pehl
}
