% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline.data.frame.R
\name{inline.data.frame}
\alias{inline.data.frame}
\title{Convert a vector of stings to a data frame}
\usage{
inline.data.frame(..., sep = ";", quote = "")
}
\arguments{
\item{...}{string, or a vector of strings}

\item{sep}{Item separator within strings, defaults to ";"}

\item{quote}{Quote character for masking separators, empty by default}
}
\value{
a data frame
}
\description{
\code{inline.data.frame()} converts a vector of strings that contain
separated items into a data frame.
}
\examples{
inline.data.frame(
    "letters; numbers",
    "A;       1",
    "B;       2",
    NULL)   # this last line allows for easy switching of line order
inline.data.frame(c("letters; numbers", "A; 1", "B; 2"))

}
\author{
Michaja Pehl
}
