% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOyear.R
\name{make.ISOyear}
\alias{make.ISOyear}
\title{speedily converting years to POSIXct values}
\usage{
make.ISOyear(years = seq(from = 1900, to = 2200, by = 5))
}
\arguments{
\item{years}{vector of year numbers to initialise closure with; defaults to
1900 to 2200 in five year steps}
}
\value{
function taking an integer year and returning a POSIXct value, based
        on July 2, 12am GMT as the middle of the year
}
\description{
\code{make.ISOyear()} creates a closure for speedily converting years to
POSIXct values. Initialise the closure with year values you know you will
need. Additional values will be added automatically.
}
\examples{
     ISOyear <- make.ISOyear()
     ISOyear(c(2005, 2010, 2100, 1900))

}
\author{
Michaja Pehl
}
