% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.gdx.R
\name{read.gdx}
\alias{read.gdx}
\title{Read item from \code{.gdx} file as quitte data frame}
\usage{
read.gdx(
  gdxName,
  requestList.name,
  fields = "l",
  colNames = NULL,
  factors = deprecated(),
  squeeze = FALSE
)
}
\arguments{
\item{gdxName}{Path to a \code{.gdx} file.}

\item{requestList.name}{Name of the item to read.}

\item{fields}{Fields to read from variables and equations.  Any of \code{l}, \code{m},
\code{lo}, \code{up}, \code{s} (or the long forms \code{level}, \code{marginal}, \code{lower}, \code{upper},
and \code{scale}).  \code{all} will return all fields.  Ignored when reading sets or
parameters.}

\item{colNames}{String vector of column names to override dimension and field
names.}

\item{factors}{Deprecated.  Do not use any more.}

\item{squeeze}{If \code{TRUE}, squeeze out any zero or EPS stored in the GDX
container.  Defaults to \code{FALSE}, i.e. all stored values (including
zeros and EPS) are returned.}
}
\value{
A quitte data frame.
}
\description{
\code{read.gdx()} is a wrapper function for \code{\link[gamstransfer:readGDX]{gamstransfer::readGDX()}} that
returns a quitte data frame.
}
\author{
Michaja Pehl
}
