% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.gdx.R
\name{read.gdx}
\alias{read.gdx}
\title{Read item from \code{.gdx} file as quitte data frame}
\usage{
read.gdx(
  gdxName,
  requestList.name,
  fields = "l",
  colNames = NULL,
  factors = deprecated(),
  squeeze = TRUE
)
}
\arguments{
\item{gdxName}{Path to a \code{.gdx} file.}

\item{requestList.name}{Name of the item to read.}

\item{fields}{Fields to read from variables and equations.  When using
\href{gdxrrw-package}{gdxrrw}, any of \code{l}, \code{m}, \code{lo}, \code{up}, \code{s}.  When using
using \href{gamstransfer-package}{gamstransfer}, \code{level}, \code{marginal}, \code{lower},
\code{upper}, and \code{scale} are understood as well.  \code{all} will return all
fields.  Ignored when reading sets or parameters.}

\item{colNames}{String vector of column names to override dimension and field
names.}

\item{factors}{Deprecated.  Do not use any more.}

\item{squeeze}{If \code{TRUE} (the default), squeeze out any zero or EPS stored in
the GDX container.}
}
\value{
A quitte data frame.
}
\description{
\code{read.gdx()} is a wrapper function for either \code{\link[gdxrrw:rgdx]{gdxrrw::rgdx()}} or
\code{\link[gamstransfer:readGDX]{gamstransfer::readGDX()}} that returns a quitte data frame.
}
\details{
\code{read.gdx()} will use \code{\link[gdxrrw:rgdx]{gdxrrw::rgdx()}} if \href{gdxrrw-package}{gdxrrw} is
installed and the option \code{quitte_force_gamstransfer} is not \code{TRUE}, otherwise
it will use \code{\link[gamstransfer:readGDX]{gamstransfer::readGDX()}}.
}
\author{
Michaja Pehl
}
