% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.gdx.R
\name{read.gdx}
\alias{read.gdx}
\title{Read item from \code{.gdx} file as quitte data frame}
\usage{
read.gdx(
  gdxName,
  requestList.name,
  fields = "l",
  colNames = NULL,
  factors = TRUE,
  squeeze = TRUE
)
}
\arguments{
\item{gdxName}{path to \code{.gdx} file}

\item{requestList.name}{name of item to read}

\item{fields}{fields to read from variables and equations (any of \code{lo},
\code{l}, \code{m}, and \code{up}); ignored for parameters}

\item{colNames}{string vector of column names to override dimension names}

\item{factors}{return non-numerical columns as factors (default) or character
vectors}

\item{squeeze}{if TRUE, squeeze out any zero or EPS stored in the GDX
container}
}
\value{
quitte data frame
}
\description{
\code{read.gdx()} is a wrapper function for \code{\link[gdxrrw:rgdx]{gdxrrw::rgdx()}} that
returns a quitte data frame.
}
\author{
Michaja Pehl
}
