% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.quitte.R
\name{read.quitte}
\alias{read.quitte}
\title{Read IAMC-style .csv files}
\usage{
read.quitte(file, sep = ";", quote = "", na.strings = c("UNDF", "NA",
  "N/A", "n_a"), convert.periods = FALSE, check.duplicates = TRUE)
}
\arguments{
\item{file}{Path of IAMC-style .csv file or vector of paths to read.}

\item{sep}{Column separator, defaults to ";".}

\item{quote}{Quote characters, empty by default.}

\item{na.strings}{Entries to interpret as NA; defaults to
\code{c("UNDF", "NA", "N/A", "n_a")}}

\item{convert.periods}{If \code{TRUE}, periods are converted to POSIXct. If
\code{FALSE} (the default), periods are numerical.}

\item{check.duplicates}{If \code{TRUE} a duplicates check will be performed
on the data. For time- and memory-critical applications this can be
switched off.}
}
\value{
A quitte data frame.
}
\description{
Reads IAMC-style .csv files into a quitte data frame.
}
\examples{
\dontrun{
read.quitte(c("some/data/file.mif", "some/other/data/file.mif"))
read.quitte("some/data/file.csv", sep = ",", quote = '"')
}

}
\author{
Michaja Pehl
}
