% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remind_timesteps-data.R
\docType{data}
\name{remind_timesteps}
\alias{remind_timesteps}
\title{REMIND time steps}
\usage{
remind_timesteps
}
\description{
A data frame containing the weights with which years contribute
to specific periods.
}
\author{
Michaja Pehl
}
