% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportDuplicates.R
\name{reportDuplicates}
\alias{reportDuplicates}
\title{Finds duplicates in a quitte object and warns about them per model and scenario.
Also warn if variables are identical but value or units differ.}
\usage{
reportDuplicates(mifdata)
}
\arguments{
\item{mifdata}{object that can be converted with as.quitte}
}
\value{
only the data that is duplicated. Has 0 rows if everything is fine
}
\description{
Finds duplicates in a quitte object and warns about them per model and scenario.
Also warn if variables are identical but value or units differ.
}
