% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.mif.R
\name{write.mif}
\alias{write.mif}
\alias{write.IAMCcsv}
\title{Write .mif file}
\usage{
write.mif(
  x,
  path,
  comment_header = NULL,
  comment = "#",
  append = FALSE,
  sep = ";",
  na = "NA"
)

write.IAMCcsv(
  x,
  path,
  comment_header = NULL,
  comment = "#",
  append = FALSE,
  sep = ",",
  na = ""
)
}
\arguments{
\item{x}{A \code{\link{quitte}} data frame.}

\item{path}{Path or connection to write to.}

\item{comment_header}{Comment header to be written to the \code{.mif} file.
Ignored if \code{append} is \code{TRUE}.}

\item{comment}{A character to prefix comment header lines with.  Must match
existing comment characters in file \code{path} if \code{append} is \code{TRUE}.}

\item{append}{Overwrite existing files (\code{FALSE}, default), or append to them
(\code{TRUE}).}

\item{sep}{Single character used to separate fields within a record.
Defaults to \verb{;}.}

\item{na}{String used for \code{NA} elements.  Defaults to \code{'NA'} for
\code{write.mif()} and to \code{''} for \code{write.IAMCcsv()}.}
}
\description{
A wrapper around \code{\link[readr:read_lines]{readr::write_lines}} for writing files conforming to the
\href{https://gitlab.pik-potsdam.de/rse/rsewiki/-/wikis/Model-Intercomparison-File-Format-(mif)}{\code{.mif} standard}.
}
\details{
\code{write.IAMCcsv()} uses commas as filed separators instead of semi-colons.
}
\examples{
write.mif(quitte_example_data, tempfile())

}
\author{
Michaja Pehl
}
